Up lại

+++ src/View/Tong/FormMain.java
@@ -103,7 +103,7 @@ public class FormMain extends javax.swing.JFrame {
         model = (DefaultTableModel) tableSanPham.getModel();
         model.setRowCount(0);
         for (SanPham sp : listSanPham) {
-            Object[] row = new Object[]{sp.getMaSP(), sp.getTenSP(), sp.getGiaBan(), sp.getNgaySanXuat(), sp.getTinhTrang(), sp.getHinhAnh()};
+            Object[] row = new Object[]{sp.getMaSP(), sp.getTenSP(), sp.getGiaBan(), sp.getNgaySanXuat(), sp.getTinhTrang()== 1 ? "Còn hàng" : "Hết hàng", sp.getHinhAnh()};
             model.addRow(row);
         }
         tableSanPham.setModel(model);
@@ -313,7 +313,6 @@ public class FormMain extends javax.swing.JFrame {
 
         jlbTrangChu.setBackground(new java.awt.Color(255, 255, 255));
         jlbTrangChu.setFont(new java.awt.Font("Arial", 1, 15)); // NOI18N
-        jlbTrangChu.setForeground(new java.awt.Color(0, 0, 0));
         jlbTrangChu.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Images/icons8-store-30.png"))); // NOI18N
         jlbTrangChu.setText("Quản lý\n\n\n\n cửa hàng"); // NOI18N
 
@@ -527,7 +526,6 @@ public class FormMain extends javax.swing.JFrame {
         });
 
         jlbDangXuat.setFont(new java.awt.Font("Arial", 1, 24)); // NOI18N
-        jlbDangXuat.setForeground(new java.awt.Color(0, 0, 0));
         jlbDangXuat.setText("Đăng xuất");
         jlbDangXuat.setAlignmentY(0.0F);
         jlbDangXuat.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
@@ -801,7 +799,7 @@ public class FormMain extends javax.swing.JFrame {
                 .addComponent(jScrollPaneSanPham, javax.swing.GroupLayout.PREFERRED_SIZE, 242, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(12, 12, 12)
                 .addComponent(PanelDetailSanPham, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(23, Short.MAX_VALUE))
+                .addContainerGap(38, Short.MAX_VALUE))
         );
 
         PanelView.add(PanelSanPham, "PanelSanPham");
@@ -909,7 +907,7 @@ public class FormMain extends javax.swing.JFrame {
                 .addComponent(JpnTacVuNhaCungCap, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jScrollPaneNhaCungCap, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(123, Short.MAX_VALUE))
+                .addContainerGap(138, Short.MAX_VALUE))
         );
 
         PanelView.add(PanelNhaCungCap, "PanelNhaCungCap");
@@ -1020,7 +1018,7 @@ public class FormMain extends javax.swing.JFrame {
                 .addComponent(jpnTacVuNhanVien, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jScrollPaneNhanVien, javax.swing.GroupLayout.PREFERRED_SIZE, 295, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap(138, Short.MAX_VALUE))
         );
 
         PanelView.add(PanelNhanVien, "PanelNhanVien");
@@ -1275,7 +1273,7 @@ public class FormMain extends javax.swing.JFrame {
         );
         PanelThongKeLayout.setVerticalGroup(
             PanelThongKeLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 477, Short.MAX_VALUE)
+            .addGap(0, 492, Short.MAX_VALUE)
         );
 
         PanelView.add(PanelThongKe, "PanelThongKe");
@@ -1463,25 +1461,34 @@ public class FormMain extends javax.swing.JFrame {
     }//GEN-LAST:event_btnTimKiemSanPhamActionPerformed
 
     private void btnXoaSanPhamActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnXoaSanPhamActionPerformed
-        SanPhamDAO cddao = new SanPhamDAO();
-        int selectedRow = tableSanPham.getSelectedRow();
-        int columnIndex = 1;
-        String cellValue = null;
-        if (selectedRow != -1) {
-            Object value = tableSanPham.getValueAt(selectedRow, columnIndex);
-            cellValue = value.toString();
-        }
 
-        if (cellValue != null) {
-            int x = cddao.removeSanPham(cellValue);
-            if (x > 0) {
-                JOptionPane.showMessageDialog(this, "Xóa thành công!");
-                fillDataToTableSanPham();
-            } else {
-                JOptionPane.showMessageDialog(this, "Xóa thất bại!");
+        try {
+            SanPhamDAO cddao = new SanPhamDAO();
+
+            int selectedRow = tableSanPham.getSelectedRow();
+            int columnIndex = 0;
+            String cellValue = null;
+            if (selectedRow != -1) {
+                Object value = tableSanPham.getValueAt(selectedRow, columnIndex);
+                cellValue = value.toString();
             }
-        } else {
-            JOptionPane.showMessageDialog(this, "Vui lòng chọn một hàng để xóa!");
+
+            if (cellValue != null) {
+                int confirm = JOptionPane.showConfirmDialog(null, "Bạn có chắc chắn muốn xoá nhân viên này?", "Xác nhận", JOptionPane.YES_NO_OPTION);
+                if (confirm == JOptionPane.YES_OPTION) {
+                    int x = cddao.removeSanPham(cellValue);
+                    if (x > 0) {
+                        JOptionPane.showMessageDialog(this, "Xóa thành công!");
+                        fillDataToTableSanPham();
+                    } else {
+                        JOptionPane.showMessageDialog(this, "Xóa thất bại!");
+                    }
+                } else {
+                    JOptionPane.showMessageDialog(this, "Vui lòng chọn một hàng để xóa!");
+                }
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }//GEN-LAST:event_btnXoaSanPhamActionPerformed
 
@@ -1505,7 +1512,7 @@ public class FormMain extends javax.swing.JFrame {
                     LabelDeTailTenSP.setText(sp.getTenSP());
                     LabelDeTailGiaBan.setText(String.valueOf(sp.getGiaBan()));
                     LabelDeTailNgaySanXuat.setText(sp.getNgaySanXuat().toString());
-                    LabelDeTailTinhTrang.setText(sp.getTinhTrang() == 1 ? "Mới" : "Cũ");
+                    LabelDeTailTinhTrang.setText(sp.getTinhTrang() == 1 ? "Còn hàng" : "Hết hàng");
                     loadHinh(sp.getHinhAnh());
                 } else {
                     JOptionPane.showMessageDialog(this, "Không tìm thấy sản phẩm với ID: " + id, "Lỗi", JOptionPane.ERROR_MESSAGE);
